[Go] consolidate List array
